feat: shared events, sweep replay protection, storage audit, sdk pin docs
- Move event struct schemas into bridgelet-shared (shared/events.rs) and
  re-export them so the SDK and both contracts share identical definitions.
- Bind the sweep authorization signature to the ephemeral account address so a
  captured signature cannot be replayed against a different account.
- Document the SweepController storage-key collision audit (all keys are
  distinct DataKey variants; Soroban isolates storage per contract instance).
- Document the centralised soroban-sdk pin and its testnet-gated upgrade
  process.
