+++ contracts/shared/src/lib.rs
@@ -1,7 +1,12 @@
 #![no_std]
 
+mod events;
 mod interfaces;
 mod types;
 
+pub use events::{
+    AccountCreated, AccountExpired, MultiPaymentReceived, PaymentReceived, ReserveReclaimed,
+    SweepExecutedMulti,
+};
 pub use interfaces::{EphemeralAccountInterface, SweepControllerInterface};
 pub use types::{AccountInfo, AccountInitRequest, AccountInitResult, AccountStatus, Payment};
