+++ contracts/sweep_controller/src/authorization.rs
@@ -4,24 +4,41 @@ use soroban_sdk::{xdr::ToXdr, Address, BytesN, Env};
 
 /// Construct the message to be signed for sweep authorization
 ///
-/// Message format: hash(destination + nonce + contract_id)
+/// Message format: hash(account + destination + nonce + contract_id)
+///
+/// Binding the ephemeral `account` into the signed payload (alongside the
+/// per-controller `nonce` and the `contract_id`) provides replay protection:
+/// a captured signature cannot be replayed against a different ephemeral
+/// account, a different destination, a different controller, or at a later
+/// time once the nonce has advanced. (#29)
 ///
 /// # Arguments
 /// * `env` - Soroban environment
+/// * `account` - Ephemeral account the sweep authorizes funds to leave
 /// * `destination` - Destination wallet address
 /// * `contract_id` - The sweep controller contract address
 ///
 /// # Returns
 /// BytesN<32> containing the hash of the message components
-fn construct_sweep_message(env: &Env, destination: &Address, contract_id: &Address) -> BytesN<32> {
+fn construct_sweep_message(
+    env: &Env,
+    account: &Address,
+    destination: &Address,
+    contract_id: &Address,
+) -> BytesN<32> {
     // Get current nonce
     let nonce = storage::get_sweep_nonce(env);
 
     // Construct the message by concatenating:
+    // - account (serialized as bytes)
     // - destination (serialized as bytes)
     // - nonce (as u64, 8 bytes)
     // - contract_id (serialized as bytes)
     let mut message = soroban_sdk::Bytes::new(env);
+
+    // Add ephemeral account address bytes so the signature is bound to it
+    let account_bytes = account.to_xdr(env);
+    message.append(&account_bytes);
 
     // Add destination address bytes
     let dest_bytes = destination.to_xdr(env);
@@ -60,7 +77,7 @@ fn construct_sweep_message(env: &Env, destination: &Address, contract_id: &Addre
 /// Ok(()) if signature is valid, Error otherwise
 pub fn verify_sweep_auth(
     env: &Env,
-    _account: &Address,
+    account: &Address,
     destination: &Address,
     signature: &BytesN<64>,
 ) -> Result<(), Error> {
@@ -71,8 +88,9 @@ pub fn verify_sweep_auth(
     // Get the sweep controller contract address
     let contract_id = env.current_contract_address();
 
-    // Construct the message that should have been signed
-    let message = construct_sweep_message(env, destination, &contract_id);
+    // Construct the message that should have been signed. The account is
+    // included so the signature cannot be replayed on a different account.
+    let message = construct_sweep_message(env, account, destination, &contract_id);
 
     // Verify the Ed25519 signature
     env.crypto()
